# storage/api-docs/api-docs-admin.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Admin application",
        "description": "API endpoints for the admin application.",
        "contact": {
            "name": "viracoding",
            "email": "viracodingGplus@gmail.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/admin",
            "description": "development Server"
        }
    ],
    "paths": {
...
        "/sectioncourses": {
            "get": {
                "tags": [
                    "SectionCourse"
                ],
                "summary": "Get sections of courses with pagination",
                "operationId": "ea1a02c51180351447586e72a3d935b4",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SectionCourse"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SectionCourse"
                ],
                "summary": "Store a new section course",
                "operationId": "4aa28fda37aad60eb53cf01cbc6da042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id",
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "course_id": {
                                        "description": "The ID of course",
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    },
                                    "title": {
                                        "description": "The title of the section",
                                        "type": "string",
                                        "example": "Section Title"
                                    },
                                    "description": {
                                        "description": "The description of the section",
                                        "type": "string",
                                        "example": "Section Description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "integer",
                                        "example": 50000
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "integer",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Section course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SectionCourse"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Section course created successfully"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/components/schemas/ValidationError"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "description": "URL to the first page",
                        "type": "string"
                    },
                    "last": {
                        "description": "URL to the last page",
                        "type": "string"
                    },
                    "prev": {
                        "description": "URL to the previous page",
                        "type": "string"
                    },
                    "next": {
                        "description": "URL to the next page",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "description": "Current page number",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Index of the first item in the current page",
                        "type": "integer"
                    },
                    "last_page": {
                        "description": "Total number of pages",
                        "type": "integer"
                    },
                    "path": {
                        "description": "URL path for the current page",
                        "type": "string"
                    },
                    "per_page": {
                        "description": "Number of items per page",
                        "type": "integer"
                    },
                    "to": {
                        "description": "Index of the last item in the current page",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Total number of items",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "ValidationError",
                "description": "خطای اعتبارسنجی",
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "status": {
                        "type": "integer",
                        "example": 404
                    },
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "دسته والد وجود ندارد!"
                    },
                    "errors": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "SectionCourse": {
                "title": "SectionCourse",
                "description": "Schema for a section of a course",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the section",
                        "type": "integer",
                        "format": "int64"
                    },
                    "course_id": {
                        "description": "The ID of the course",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "The title of the session",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the course.",
                        "type": "number"
                    },
                    "discount": {
                        "description": "The discount applied to the course.",
                        "type": "number"
                    },
                    "description": {
                        "description": "The description of the session",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The datetime when the session was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The datetime when the session was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}